requirements.txt
pocketgroq==0.3.0
python-docx==1.1.0
python-dotenv==1.0.0
bs4